Refactor app boot logic
